Доработки к уроку 1
Разделила компоненты по отдельным файлам.
