+++ Anna_Tokareva/lesson_1/src/components/Message.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+
+export const Message = ({ name, content }) => (
+  <div>
+    <strong>{name}: </strong>
+    {content}
+  </div>
+);
