+++ Anna_Tokareva/lesson_1/webpack.config.js
@@ -27,5 +27,8 @@ module.exports = {
         }
       }
     ]
+  },
+  resolve: {
+    extensions: [".jsx", ".js"]
   }
 };
